strategy-layer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "strategy-layer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "redis[asyncio]>=5.0",
    "pydantic>=2.5",
    "watchdog>=3.0",
    "structlog>=23.0",
    "numba>=0.58",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["strategy_layer*"]
